Timers and object spawner component
-Timers added:
 Timers are an object that counts down from some number to 0,
decrementing by 1 each second.  When the timer runs out, some action is
performed, and the timer may be recreated again. While Timers are in the
World file, they do not actually have a location in the world, and
instead reside on a master list the server keeps track of and updates
each game loop.
-object spawner component added:
 The object spawner component is placed in the world, attached to some
object, after the object has been placed.  It stays in the room it was
placed in regardless if the item it is attached to is picked up or
moved, and will make timers on a regular basis that do a random check to
generate the item again in the room if the check is passed. Object
spawners can also be set to repeat, in which case copies of the object
the spawner is attached to will be periodically be created in the room.

Note: I am trying to be more detailed with my commit messages, to better explain the things I am adding.  Also, this particular commit has brought to my attention a bug where 'look' and 'examine' commands show the description for all items in the room with the same name.  These commands should only ever show the description of one item at a time.