'look' and 'examine' commands have indexing
-'look' and 'examine' commands now allow indexing.  When there are
multiple items in a room with the same name, sending the usual command
for look or examine will always display the first object listed.  To see
the other objects, add a number after the object name in the command,
where the number corresponds to the objects position in the list
displayed.  For example, to look at the second cat in a room with 2
cats, the command would be 'look cat 2'.  Using an index when there is
only 1 object of the given name will result in no found objects, so
remember to remit the index if there is only 1 item.