'get' indexing, locked containers, prettier and more informative inventory, inventory limit
-'get' command now has indexing, which works the same as it does for
'look' and 'examine'. When using indexing, it is always best to use the
item's proper name, with '_' for spaces in the name.
-containers can now be set to locked.  When locked, the contents of a
container are invisible and are not interactable.  Players do see that a
container is locked when they examine it.
-inventory display is cleaned up to be more readable.
-inventory now has a size limit, which is just a paramater of the
clients' avatars.  This means that it will later be possible to increase
max inventory size through items or equipment.
-various bug fixes, especially with indexing