Password authentication for users, beginning of server data persistence
-Players now must authenticate with a password when logging in.  The
first time a name is used, the player is prompted for a password.
Future login attempts, even after the server has been restarted, will
require the player to enter the password they set on the first attempt.
  Failure to do so 3 times in a row kicks the player, although they are
free to connect and try again.
-The beginnings of a framework for saving and loading the data on the
server when the server is shutdown and restarted.  This system will also
save and load player data when players log on and off.
-passlib added to handle hashing for passwords