objects can now be 'notDroppable', and objects in a player's inventory now correctly change their currentRoom when the player moves around
-Objects that have 'notDroppable' = True will not be able to be dropped
after they are picked up.  This will generally be reserved for things
like keys
-Objects being held by a player as that player moves around now adjust
their currentRoom attribute to reflect the room the player is in
-I suspect a bug related to newly spawned objects just being references
to the object the objectSpawner is attached to, rather than being
separate in memory.  If this is the case, I need to first make a generic
object, and then add all the components and such to make it a copy of
the parent object, and then place the newly formed object in the correct
room.  Will look into this for future commit.