Moving towards persistance and a file-based object structure
-players can now be 'OP' with special commands available to only them.
-'spawn' command added for OP players. Allows the creation of objects in
the world.  Currently, the ability to spawn objects inside containers is
broken.
-objects are now defined by files, which live in the 'obj' directory and
are loaded at server start.  Eventually this will aid in server
persistance and the arbitrary modification of object stats.
-a half-finished 'save' functionality for players has been written.
Currently, it outputs (mostly) correct player files, but there is no
mechanism to load those files in place yet.

I'm going to continue shifting the world over to being stored in files
and so that the entire server can be saved/loaded on startup/shutdown,
and player's data can be loaded in on connect and loaded out on
disconnect.