World is now file-based
-Big overhaul of the world, and how it is structured.  The entire thing
now is described in human-readable text files that load into the server
at startup.  When the world is built and rooms and items are created,
new copies of the objects described in these files are created, with new
memory addresses.  This should *greatly* foster easy content creation.
-objects cannot yet be put in inventories since the switch.  This should
be fixed soon.
-'get' command with objects that are in inventories is still broken.
Claiming objects from inventories will instead be moved to a 'check'
command that automatically grabs everything in an inventory, or if that
would be too much for the player to carry, grabs the first items listed
until the player's inventory is full.
-Overall, files seem more stable, but it is possible there could be some
residual bugginess appear in the near future as a result of the big
switch.  In any case, the march towards being able to save and load the
server and player profiles carries on.  Once everything is working
properly with content described in text files, saving and loading the
world is simply a matter of creating and reading those text files.