Room Persistance mostly working
-When the server receives a 'shutdown' command, it now saves all the
rooms in the world, including who and what is in them
-If the world has been saved and exists in 'data/world', the world will
load from that save file on server start instead of the blueprint
-Initial rooms and item definitions moved to 'blueprints' folder
-Still need to add in output to save file for objects that spawn in
rooms other than the one they are located in, and output to save file
for objects that spawn in containers in rooms other than the one they
are located in.
-Still lots of debug mess in the server's console output...hopefully
will be cleaned up for next commit.