Bugfixes, mainly for get and item loading
-It appears the bugs mentioned in the previous commit are all fixed at
this point(however, see notes below).
-When objects are spawned into the world with the entrance of a new
player, they now only generate one server message, as intended
-Objects spawning into the world with the entrance of a new player do
not alert their arrival to others in the room, as intended
-'get' now only grabs one item from containers, not two, as intended
-items being spawned into the world by an objectSpawner into a container
no longer duplicate themselves after a short time.  They now only ever
spawn once per check, as intended.
-it appears that get now always successfully removes the item from the
room.  However, exactly once this was not the case, and a "garden_gnome"
in the "test:lobby" got stuck as per the old bug.  I was not running a
debugger at the time, and I have not been able to reproduce this error.
  For now I'm saying that this is fixed.
-there is once again more debug-related prints in the server console.  I
swear, this junk will one day be gone!