bugfix:using 'get' on objectSpawner-created objects
-Finally tracked down and fixed bug where sometimes items would not be
removed from a room when 'get' was used.  It turns out the
itemGrabHandler component for each new object was still pointing to the
template itemGrabHandler, instead of their own.  Because of this, when
the itemGrabHandler attempted to remove it's owner.owner from the room,
it wasn't found, because the template object is never in a room