'check' command mostly working
-'check' command implemented.  'check' gives a player every object in a
container, assuming the container is unlocked.  'check' also allows
specifying a container when there are duplicates in a room (well, will
eventually.  This is buggy right now)
-identically named containers in the same room cause problems for both
'get' and 'check' right now.  In a future commit, 'get' will only work
for objects in the room, whereas 'check' will work for objects in
containers only.