'get' working, 'check' added
-'get' command now works as intended.  Furthermore, 'get' indexing is
different.  Now, the index searches all items of the same name in the
room, in both containers and out in the open.  Thus, "get rock 2" would
get the second rock in the room, regardless of what that rock is inside
of.
-'check' command added.  Check gets everything in a container and puts
it in a player's inventory, if they have space.  Otherwise, they get
what they have space for
-stuff lists now have a form of indexing when naming containers.  The
items always go in order, but now there is a way to specify which of
several identical containers in a room an item should be created in.
Entries in the stuffList for a room now look like: "rock:trash_can1"
-items that spawn in containers no longer alert on their arrival
-cleaned up debug print messages in the console
-the " ' " command now switches between telling the last player to send
a tell to a client, and the main chat channel, depending on what was
last used.