Mob spawners working
- The mob spawner component now exists.  It is added on to a base Object
to create a mob spawner, which is an invisible object in a room that
causes the periodic creation of mobs.  It has a timer that defines how
often to check for spawn events, and a mode that dictates spawn
behavior.  When the mode is 'cont', each successful spawn event creates
one mob in the room.  When the mode is 'thresh', a successful spawn
event spawns a mob, but only if the number of mobs in the room is less
than the 'cycles' parameter.  When the mode is 'pack', each spawn event
creates a number of mobs in the room, up to the number defined by
'cycles'.
-mob spawner objects are persistent in the world between server reboots