mobs now expire
-mobs expire after a set amount of time if they do not have contact with
a player.  Right now, when a player moves into the room a mob is in,
every mob in the room resets their expiration timer.  In the future,
when a mob moves into a room a player is in, this will also reset the
mob's expiration timer.  However, if a mob and a player are sharing a
room, after the initial expiration timer reset, the timer will continue
to tick down and expire the mob.  This means that afk players in a room
do NOT prevent a mob from despawning due to lack of player contact.
-There is a bug right now where existing players' inventories spawn both
on the player and in the world when the player reconnects to the server.
  Will fix this soon.