Beginning of battle system
- 'fight' command implemented.  'fight' starts a battle between a player
and a mob.  When a battle starts, both the player and the mob are moved
into a temporary 'battle room', and a marker of the battle is left in
the room the player was in when the battle was initiated.
-'flee' command implemented.  'flee' only works in a battle room.  When
a player 'flee's, all the contents of the battle room are placed back in
the original room, and the temporary battle room is destroyed.
- currently, there seems to be a bug with roomFile-defined mobs in a
battle after restarting the server.  In the off chance that a player is
in battle with a pre-defined mob when the server shuts down, when the
server restarts and the player reconnects, the player will spawn into an
empty room, with the pre-defined mob gone completely.  Since pre-defined
mobs is not a good practice (even though it is possible when designing
rooms), this is a minor bug.