Basic battle system bugs squashed, battle system display cleaned up, new region added
- The very basic battle system in place no longer has bugs associated
with the end of battle.  At this point, the only remaining bug I am
aware of happens when a player has reward money or reward exp and the
server shuts down and reloads.  In that case, the reward money and
reward exp will disappear, because I have not yet coded the routine to
save and restore them
- the battle system is now a bit easier to read, and more informative
- a new region, 'testb' or test basement, has been added.  It contains a
new mob, a spider, and a new mob spawner, spider_web.  This was to test
that I can easily make new regions, rooms, mobs, and spawners, and that
they would work properly.  Also, I have confirmed that mobs do not leave
the region they are spawned in, as intended.