Exp added to status, colorized hp and pp in status, added numerical movement system
-'status' command now shows total exp
-the colors for the hp and pp in the 'status' screen now change to
reflect condition
-numerical movement now works.  Sending just the number of the exit
(corresponding to it's place on the list displayed) will follow the exit
as if you had typed it's full name out.  This should be compatible with
movement macros, and make movement much easier.
-added the very simple aiBattle script, which for now is just a testing
placeholder for the actual mob battle ai.  It should have been added in
an earlier commit, but I just now noticed it hadn't been added.