Fixed bugs relating to battles
-Fixed a number of bugs related to ending battles, as well as bugs
related to shutting down and restarting the server while a battle was
happening.  Battles seem to be pretty robust at this point.
-There is a potential memory leak related to shutting down and
restarting the server while a battle is happening.  When the server
shuts down, any ongoing battles are saved to a file like other rooms.
When the server restarts, it re-creates the world from the saved room
files, including the saved battle room.  It does this by creating a
prototype room from the file definition, and then spawning a new room
identical to the prototype.  Unfortunately, right now the prototype is
not deleted...so every battle that is happening when a server shuts down
ultimately creates a useless file, and a useless object in the game.
This will be fixed in the next commit.
-There is a bug where a player sees the wrong display while in a battle
room that was created because the server just restarted.  It should be
showing a battle display, but it is instead showing the regular room
display, which includes the exits and does not show health.