Leveling Up, ExpToLevel in status
-'status' now shows the exp required before reaching the next level
-when a player's TNL as displayed by 'status' reaches 0, the player now
gains a level.  A number of options for stat boosts are displayed, and
the player selects one to gain those stat boosts.  The player is unable
to move or initiate combat until a selection is made, as the played is
switched to the 'levelup' game state until a choice is made.