onUse for items working, with a few examples made
items can now have an onUse function attached to them.  When a player
uses an item, the item will attempt to run it's onUse function, if it
has one.  This function then determines the effect of the item.  Both
healing and causing damage to a mob while in battle are already coded,
with the flexibility to add MANY more of these effects in the future.