Beginning of Equipment, uses.py added (forgot last commit)
Equipment is starting to take shape.  Right now, equipment is another
component added to an item, which then supplies attributes to the
player.  Equipment lives in a separate list in each room.  So far,
equipment can be defined in files and spawned into a fresh world.
Players can also pick it up and drop it.  It currently is not equipable,
and does not affect player stats.
Also, uses.py was forgotten last commit, so here it is.