Mostly Concerning File creation and Loading around Equipment
-equipment that is located in rooms is just about functional.  I still
need to handle loading equipment into the world from the saved files, as
right now it is ONLY loading in template copies based on what is defined
in the room.  It should default to loading based on what files are
present, instead.
-equipment saving and loading is not yet implemented for equipment on
the player, and equipment on mobs.  I'm pretty sure equipment isn't even
being saved when held in a player's inventory, rather than equipped