Duplicate Login Handling
-it is now no longer possible for two clients to be logged in as one
username.  If a client successfully enters the password and logs into a
username that is already logged in, the oldest client is kicked.  The
assumption is that clients won't be able to kick each other and abuse
this unless they already have the password required to log in to the
account.
-when a player is kicked by a duplicate login, unfortunately they do not
see a message explaining what happened, but instead just lose input and
feedback from the server.  The server ends the session before the
message is able to propagate to the client, and so the client never
recieves it.  For now though, the functionality is there, but players
may be frustrated when they don't understand why the MUD seems to lock
up on them.