Equipment is equipable, 'equip' command added, inventory display updated
- equipment can now be equipped.  Each bit of gear has a slot that it
will equip to.  Each slot is unique, but you can have gear equipped to
as many slots as you'd like, as long as none are the same.  Equipped
gear adds stats directly to your character.  The status screen now
reflects your actual and base value for each stat, where the actual is
determined by the base value summed with the bonuses from equipped gear.
-the 'equip' command is now active, which equips a bit of equipment that
is currently in a player's inventory
-the inventory display has been updated to be visually consistent with
the status display
-the inventory display now shows when something is equipped
-the description for weapons and armor has been modified in the
inventory display to be more useful
-right now, it is not possible to remove equipment once it has been
equipped.  It is probably possible to drop it, though.  That needs to be
fixed asap.