Added unit test framework and a few brief examples
-It is terrible I went this long without any sort of testing.  I have
setup test.py and mock in order to be able to write tests without
boilerplate.  These tests are now located in the 'unittest' folder.
Hopefully this will encourage me to actually writing some tests when I
run into bugs in order to begin remedying my lack of testing thus far.
In other words, this is a guilt commit, but should make the project
better in the long run.