onExit attribute added to rooms
-rooms can now have an onExit dictionary, which is a dictionary defining
a function that runs when the player takes a given exit.  This allows
code to be run, based on where the player is, that interrupts movement
and causes something to happen, and then either continuing the move to
the room the player wants, or teleporting elsewhere
-two very useful onExits added: teleportRandom and gotoRoom.
teleportRandom takes in a region, and moves the player to a random room
in that region.  gotoRoom takes in a region and a room, and moves the
player directly to the room.
-added onExits.py, to define onExit functions